move ResolveUniquePath to OS specific files
also adds filepath.Clean() to end of function
